jalal_partie/pyproject.toml
[tool.poetry]
name = "cluster_reduc_data2"
version = "0.1.0"
description = ""
authors = ["khaldi <jalalkhaldi3@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
pandas = "^2.0.0"
numpy = "^1.24.2"
scikit-learn = "^1.2.2"
matplotlib = "^3.7.1"
nltk = "^3.8.1"
emoji = "^2.2.0"
gensim = "^4.3.1"
spacy = "^3.5.2"
torch = "^2.0.0"
umap-learn = "^0.5.3"
coclust = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
